Include Dia schematic and extend user signup
The systems diagram is now created with Dia, although it's not complete
yet.  There are also some small modifications in the Makefile and
included build.mk files.

The program was extended by including further steps in the signup
process for a new user.  The client-dedicated process packs all relevant
user information into a single character buffer.  Fields in the buffer
are separated by the null-character.