Complete new user signup process
The signup process for the new user is mostly complete.  Or at least its
core part.  The chosen username is checked against the set of existing
users and if not taken, the new user is added to the vector of users
residing in memory, while their details (userID, personal name, username
and encrypted password) are written into the users file (nchatUsers).