Complete login functionality at server
The core of login functionality at the server is completed although it
hasn't been debugged yet.  The distributor process is now able to take
the message containing the username and password from the
client-dedicated process, unpack it and check if the username exists in
the database and if the sent password matches the password stored there
(including prepending salt).  It sends back the message to the
client-dedicated process: OK for username found and password matched,
NOK for password that doesn't match, and UNK when the username hasn't
been found.