Complete login functionality; start adding logout
The loging functionality has been mostly implemented and debugged on the
server side.  The client side is capable of receiving an appropriate
response from the server, but so far does nothing with it other than
opening the signup dialog only for debugging purposes.

The logout process implementation has started on the server side, but is
so far very rudimentary and inefficient.  It will require at least some
debugging and rethinking of some parts of the process.  In particular,
informing contacts that a user has logged out is for the moment very
inefficient as it involved a double loop through the list of all users.