Ignore child process termination at server
The sigaction() call is used in the listening process in order to start
ignoring the SIGCHLD signals sent when a client-dedicated process
terminates.  This is done in order to prevent leaving client-dedicated
processes in the zombie-state when they are finished.

Ideally, this should be done through a handler that may potentially need
to do some cleaning up (for example, by informing the distributor
process to change user state if a user has been logged on through the
terminated client-dedicated process).  Preliminary implementation of
such a handler didn't work well and is left for later.  It seemed that
when this handler was active, new client-dedicated processes were
spawned when the original one was terminated.  It could have been due to
incorrect implementation of the wait() function, but this hasn't been
thoroughly tested.