Improve message processing at distributor process
Messages were previously entirely read at the distributor process and
only taken off the queue if they are meant for it.  This leads to more
processing, but also to a subtle error when a distributor acts on a
message that is not meant for it.  This can cause the queue to overfill
if the distributor acts on mCheckUser message sent to the
client-dedicated process when the distributor process starts processing
it first.  It doesn't remove it from the queue, but gets tricked into
thinking that it needs to check the user again and quickly starts
sending more messages to itself.  The problem seems to be solved by
instructing the distributor process to process only messages where it is
specified as the recepient.