Start implementing main client window
This contains the start of the implementation of the main window that
should open when a user successfully logs in or after it signs up.  So
far, only a rudimentary class for the top menu has been implemented (but
not tested), and there's a start of the background panel on top of which
other elements should be positioned: the top menu, the left-hand-side
panel with the list of contacts, and two larger panels on the
right-hand-side: current message at the bottom and message history at
the top.