Start coding the sending of contact list
The background panel of the client application is now split into three
sections, one of which will contain a list of contacts for the logged on
user.  This list will be populated when the server sends the list of
contacts.  However, this will require some redesign.  At the moment, the
client-dedicated process goes through an infinite loop in which it
receives messages from a client.  It can exchange messages with the
distributor process, but only in response to messages received from the
connected client.  The problem there is that the distributor process may
need to send some unsolicited messages to the client-dedicated process,
which the client-dedicated process may never receive as it is occupied
with waiting for messages from the client.

A possible way out of this is to set the socket for client connection
into the non-blocking mode.  The loop could then first go through
possible client messages, and then turn towards possible distributor
process messages within the same iteration.  This problem has to be
solved before clients start exchanging messages with each other.  Even
if sending of the contact list by distributor can be solved with current
implementation (as contact list is deterministically send whenever a new
user logs in), there is no way for the current implementation to deal
with asynchronous events, such as a contact logging in/out or sending a
message to the user.