Start redesigning server
It was a mistake to deal with incoming connection requests by spawning
child processes.  That leads to a lot of problems with communicating
with them using shared memory.  It seems that blocking on an accept()
call for a socket can be solved using the select() function that will
continue the execution if any socket is ready for reading or if the
listening socket is ready to accept a new connection.