Replace multiple server processes with single one
This solution involves the select() function and appears to be much more
elegant than splitting the server into the listening, distributor and
client-dedicated processes.  It avoids inter-process communication which
should speed processing up and avoid potential deadlocks.

This has only been implemented at the server side for the moment.  And
it hasn't been debugged.