Consolidate socket send and recv functions
This continues the work on streamlining sending and receiving with
sockets.  Several calls on the process side are replaced with calls that
include a new Process class function for converting a buffer of
characters into a set of strings.

In addition, several const Socket* pointers are converted to just
Socket* as some of the recv calls modify Socket objects by resizing the
buffer, and couldn't be  associated with const Socket* pointers.