Clean up messages and processing functions
Some messages exchanged between the server and the client and the
corresponding server-side functions are renamed.  There were also a few
changes to some socket functions.  Short functions for sending simple
data types are simplified and moved to the header file.

The login button in the login dialog now again properly processes the
success response from the server.  The signup dialog hasn't been updated
yet to deal with the new server setup.