Complete signup with new server implementation
The signup function has been redesigned to be compatible with the new
implementation of the server that isn't split into several processes any
more.  This made the process of checking if the chosen username is
available and subsequent adding of the new user more streamlined.  The
MessageType has also been slightly reorganized to accommodate for these
changes.