Implement main window top menu buttons
Only two menu entries are implemented so far: for interaction with
contacts and for the account settings.  Selecting each entry opens
another menu.  The Contacts entry opens a new menu with a single entry:
searching for a new contact.  The Accounts entry opens a new menu with
two entries: log out or quit.  The actions of these two have been
partially implemented, but not tested yet.