Start implementing contact requests
The entries of the top menu in the main client window are largely in
place even though their functionality is not implemented yet.  This
process starts by working on the Find contacts entry in the Contacts
menu.  At the moment, activating this entry opens a new dialog in which
the user can enter the username of a requested contact.  This is then
sent to the server.

The server process goes through the list of users and tries to find the
requested contact.  This requires a new function findUser whose
implementation has started.

If the requested user has been found, but is offline, the connection
request will have to be logged into its User object (and written into
the users' file).  A new member is added to the User object to store
this information.  It's a vector of integers, where each entry is the
user ID of a user who requested the contact.