Continue implementing contact request
A simple Dialog-inherited class is created to display a message to the
user if their contact has been found on the server or not.  It should be
shown when the response from the server about the search status has been
received, which should normally happen immediately after the contact
request has been sent.

Reception of server response to contact request at the client side is
not implemented yet.  One option is to block the program until this
response is received as it should arrive immediately after the contact
request is sent.  But another option could be to receive is
asynchronously as the server could send the name of the requested
contact so that the client doesn't have to keep this information and
wait for the server response at the point of sending the request.