Start implementing select() on client
The select() function implementation is started on the client side.  The
idea is to block the execution until there is either a keyboard event or
data becomes available for reading on the socket.  This call takes place
in the NCWindow::run() function as the central function of the
application after a user has logged in.