Test select() on handling TopMenu
The select() function seems to be working as expected inside the
NCWindow::run() function, at least in its handling of the top menu.
Instead of using the NCursesMenu::operator() which creates a while loop
until the menu is quit, the select() function is placed in the
NCWindow::run() and if the appropriate file descriptor (stdin) is ready,
the TopMenu::handleKey() function is invoked to accept the key and call
the menu driver.

This will all have to be handled in a better way when other panels in
the NCWindow are implemented as the focus will not always be on the top
menu.  A possible way would be to use virtual functions and store
pointers to all relevant elements in a vector.  All of these elements
are derived from NCursesPanel, but the panel itself doesn't have a
corresponding handleKey() function.  Perhaps its getKey() function could
be overridden for this purpose?