Update FD_SET() for select() call
The select() call modifies file descriptor sets on its exit, which
wasn't taken into account.  At the arrival of the first message on the
client-oriented socket, the bit for the listening socket at the server
process was cleared, which made it impossible for new connections
requests to be accepted.  This has now been fixed with calling FD_SET()
before the select() call in every loop iteration.