Implement dialog for response to finding contact
When the client sends a request to establish a contact with another
user, the server responds with whether the user has been found or not.
This response is now acknowledged by the client by displaying a pop-up
box that is an object of a class that inherits from Dialog.  It only
displays the status of the query (whether the requested contact has been
found or not) and has only an OK button.

Two things remain to be implemented with this pop-up box: sending an
error response when a user tries to establish a contact with themselves,
and fixing the focus in the pop-up box, which now seems to be on the
label rather than the OK button.  At the moment, in order for the box to
close, the user first has to press a tab and then enter.