Add sending and receiving contact requests to User
The User class is modified to include two lists of user IDs for sending
and receiving contact requests.  These are dealt with separately in
order to later use them for two different kinds of entries in the list
of all users displayed in the client window.  The server process is also
modified to populate these two separate lists.