Modify handleKey() function of TopMenu
The handleKey() function now receives the key from the outside -- when
invoked.  Previously the getKey() function was called from it.  However,
the new approach is able to do the preprocessing of the key outside,
which can be used to change the panel if a tab key has been pressed.
The goal is to be able to switch between panels (top menu, contacts,
message history and current message) by pressing the tab key.