Provide more details in Dia schematic
The initial schematic only had basic elements and their relationships.
A more detailed schematic is added below the original to show what
elements have been implemented so far.  Some of these are still missing,
such as several items in various menus.  In fact, it may be a bad design
decision to have so many different classes for various menu items, but
that seems to be approach taken by the ncurses demo developers.