Extend tracing to serverProcess.cpp
The TRACE macro is applied to the constructor and the message loop of
the server process.  This requires the verbosity level to be known
inside the server process object.  As it didn't seem practical to have
the verbosity level as a member of the server process object, it was
made as a global variable, which may not be the best choice.