Start implementing Acceptor class
The Acceptor class is implemented, but so far in a very crude way.  It
is still in the experimental stage.  An instance is created in the run()
function of the server process, and it then accepts the connection and
passes the pointer to the client-oriented socket object back to the
server process that acts as its observer.  However, the approach seen
online for the Observer design pattern hasn't been completely followed
as there seems to be only one observer for the Acceptor object -- the
server process.