Implement Acceptor class with Observer pattern
The main component of the server has been renamed from ServerProcess to
Server.  As it will not be spawning new processes, there is no need for
it to derive from the Process base class.  Instead, a new Observer class
is added and the Server derives from it.  The idea is to try an
implementation of the Observer design pattern, but I have probably done
it in a clumsy way.

A couple of the concepts I used seem dubious and some are probably
wrong.  I wanted to create a react() function of the Observer class and
make it pure virtual so that specific implementations (like Server) have
to override it.  But in order for it to be generic, I pass void *
pointer as the input argument.  The Server needs this pointer to be a
Socket *, so I used static_cast to convert it, but I'm not sure if it's
allowed in this case.

The Acceptor class acts as a publisher for the Observer.  It has only
one observer (the Server component), so it doesn't create a vector of
Observer * pointers.  But it still keeps an Observer * pointer even
though it can only be a Server * pointer.