Replace char * buffer with std::vector<char>
Local copies of char * buffers are replaced with instances of
std::vector<char>, but the buffer that is a part of Socket objects is
left in place for the moment.  It is not be used much, and it should be
either removed or replaced by std::vector<char> in future modifications.