Start implementing Connection class
The Connection objects will be created by the server when a client
connection is accepted by the Acceptor object.  The Server object
registers as an observer with Connection objects in addition to being an
observer of the acceptor.

The Acceptor and Server classes are also slightly redesigned.  The
listening socket seemed to be better placed in the acceptor and is now
removed from Server in order to be placed there.

The Acceptor object now creates Socket objects for every established
connection and passes the pointers to them to the Server, which uses
them to create Connection objects by passing Socket pointers during
their construction.  However, I doubt that I'm doing this correctly in
terms of resource management.  I will have to look into this more
carefully.  Perhaps the std::move() would help in dealing with this if I
manage to understand how it works.

One final awkward solution here is collection of socket descriptors from
the Acceptor and Connection objects in the file descriptor set in the
event loop of the Server object.  Is it correct for the Server object to
access those elements that belong to Acceptor and Connection objects?