Extend implementation of Connection class
The receive() function of the Connection class is now implemented.  It
receives data from the connected client and passes the message to the
server for further processing.

The transmit function hasn't been implemented yet.  It is envisaged to
receive a message from the server and then just transmit it to the
client.

One complication in the latest implementation is that the Observer base
class has now been modified so much that it's hardly generic any more.
It has two react() functions, one of which seems highly specialised and
receives arguments that seem to be useful only for communication with
the Server object.