Continue implementing Connection class
Most of the work on the Connection class is completed.  It can now
receive external messages and pass them to the server, and it can get
messages from the server and forward them to the client.

The implementation file has been moved from the src/server directory
into src directory with the idea that the class can be reused in the
client implementation.  That required change of the src/build.mk file.
At the same time, the other build.mk files (in src/server and src/client
directories) are modified to remove linking options that were previously
required for interprocess communication (-lrt and -lpthread).