Deactivate form in FindUserDialog
The label field in FindUserDialog (dialog that pops up as a response to
a request for establishing a contact with another user) was supposed to
be inactive, but by default any object of Dialog class or a class
derived from it starts with the form being the active part of the dialog
(selected by inForm = true and inMenu = false).  This behavior is
reversed in the constructor for the FindUserDialog because the only
field in the form is the label field that just displays the user
information and can't be selected.