Make FindContactItem in TopMenu more responsive
The default behavior for the FindContactItem was previously to stay
activated after a new contact request has been sent to the server.  That
requires the user to explicitly exit this sub-menu in order to navigate
to the root of the top menu.  This is now changed so that the default
behavior is to close this sub-menu upon issuing the contact request.
The return value of the action() function is set to true to achieve
this.