Implement logout signal sending at client
The server was already able to deal with receiving and responding to the
logout signal from the client, but the earlier implementation of the
client didn't send any signal when the logout button in the top menu was
selected.  This has now been implemented.

This implementation required several minor modifications of underlying
classes.  As the server expected the client username to be sent with the
logout message, the username is now embedded inside the NCWindow class
so that it becomes accessible to all elements created by NCWindow.  This
also requires that the pointer to the NCWindow object is embedded into
elements that it opens so that they have access to it and its members
(including the username).