Modify Socket::acceptConnection() function
The function now creates a local Socket object and returns it to the
caller by value.  Previous implementation created an object on the heap,
then returned the reference to it, which was then used in the
copy-constructor to create a new Socket object on the heap.  This meant
that the object created in acceptConnection() would never be deleted,
causing a memory leak.