Implement contact request popup button actions
The action() functions have been completed for all contact request popup
menu items.  The items are redesigned to contain a pointer to the popup
object that contains them.  An int-type member named response was
included in the Popup class in order to save the response from button
actions.

The Message class was deleted from message.h as it isn't used in the
current implementation.  It was used with the shared memory approach.