Handle response to contact request at server
The server is now able to receive a response from a client on the
contact request from another client.  If the request is accepted, the
server now updates the contact lists for both the user that sent and the
one that received the contact request.

This information is still not propagated to the user who initiated the
contact request.  And it is not written into the users file yet.