Extend response to contact request at client side
The client that received the contact request can now accept the request,
which leads to immediate placement of the established contact entry in
the list of contacts.  This information on the response of the requested
user still doesn't get sent to the user who initiated the contact
request.

A bug is removed in the Server::logout() function which was previously
sending logout information to all contacts when a user logs out.  This
doesn't work when a contact is already logged out as it leads to broken
socket pipe errors.  The message is sent to a user only if their
Connection object is not 0.