Add server functions for updating users file
The Server::updateUsersFile() function is created in order to be called
when permanent properties of a user change.  An example is when two
users establish a contact (which is implemented), or when a user changes
a password (not implemented).

A friend operator<<() has been added to the User class in order to
simplify writing of user data into a file.  For the moment, it writes
user ID, username, real name, password and the list of established
contacts, but not yet the list of initiated, but incomplete, contact
requests.