Add writing initiated contact requests to file
Contact requests that have been initiated, but not yet fully established
are written into the users file.  The file writing is also now done as
soon as a requested contact is found -- not only after the connection
has been fully established.  This should allow users to initiate a
contact and log out before the contact request has been accepted.