Add User function for self-transmission
User data is sent from the server process, but it has to be sent on
several occasions: for established contacts, for users who initiated
contact requests with another user and for users who received the
contact requests.  Instead of copying the code three times, instructions
for sending user data are now placed in a function transmit() that is a
member function of objects of User class.  That allows iterating through
the list of contacts and just calling the function for each of them.