Receive established contact from server
The client program receives information about established contacts from
the server and creates an entry in the contacts panel.  This is done in
a crude way at the moment: by parsing the received buffer with user
information field by field in the main loop of the client program.  A
better approach would be to create a member function of the User class
(or a Contact class) that will pick up this information with a single
call from the main client loop.

An enumerated type ContactType has been added to contactMenu.h in order
to make a distinction between established contacts and sent and received
contact requests.  This is mostly used when entries are created in the
contacts panel of the client program.

A minor change has also been made on the server side.  The server now
checks if another contact already exists in the list of contacts for a
user before it is pushed into the vector of sent and received contact
requests.  The same test should probably be implemented with established
contacts, but it hasn't been done yet.

Finally, as objects of the User (and perhaps Contact) class are likely
to be created at the client side, the user.cpp source has been moved to
the src directory from src/server.  This seems a more intuitive location
for files with shared code.