Implement panel switching in client program
The user that logged in can now loop through various panels in the main
window by pressing the horizontal tab key.  This allows continuing the
work on the entries for different contacts in the contacts panel.  The
ContactMenu that is used to represent an entry for an established
contact has been given a handleKey() function, which seems to be working
in its current implementation.  The function is just copied from the
TopMenu class, which indicates that a better solution here may be to
derive a new class from NCursesMenu, implement a handleKey() function in
the new class, and then use it as a base class for both TopMenu and
ContactMenu classes (and also for classes that will represent entries
for sent and received contact requests).