Start implementing text messaging on client side
The client program can now send text messages to another user, although
this feature is still very crude and in the testing phase.  The code
that processes it is still in the main loop of the client program,
whereas a better solution would be to place it within the message panel,
which will be done after the functionality is tested.

The server can receive a text message, but it can't process it yet.  It
is ignored for the time being, but it shouldn't be too difficult to just
forward it to the other user.  Implementing the display of the text
message at the end of the other user is a more complicated problem as it
will require the text to be displayed in the message history panel.  For
the time being, the decision hasn't been made on whether the text
message should be displayed as soon as it arrives (probably not) or when
the other user selects the user that sent the message in the contacts
panel (more likely, but probably more complicated as it will require
storing the message in a buffer until it's displayed).