Implement text message reception and display
This includes a preliminary implementation of forwarding of text
messages at the server, and a basic, rudimentary implementation of their
reception and display at the client side.  All text messages that are
received by the client are displayed in the message history panel.  A
better implementation would involved displaying of message history only
for a contact that is currently selected in the contacts panel, but that
will require messages to be stored in a contact-specific buffer and
possibly written into a file for offline storage.