Clear message panel after sending message
The message that has been sent is transferred from the message panel
into the history panel, where it is right-aligned.  The message panel is
cleared in the process and ready for a new message to be typed.  Typing
the new message is still an awkward process as it requires sending the
text button in the entry for the recipient in the contact list.  When a
message is sent the program doesn't keep the status indicating which
panel is selected.  That will be fixed later.

History panel and message panel are modified to serve as a background
and placeholder for inner panels that are drawn inside of them.  These
inner panels don't have any frames, which seems to allow easier handling
of text wrapping.

At the moment the panels don't handle backspace characters well.  While
the backspace deletes a previous character in the message panel, this is
not reflected in the message content, which contains both the deleted
character and the backspace itself.  Possibly, the message content can
be parsed and cleaned before sending and transferring to the history
panel.

There is also a minor change on the server side.  The server now
forwards a text message to a user only if they are online.  Otherwise,
the message is dropped.  Otherwise, the server crashes as it tries
sending content through a nonexistent connection.