Process backspace in message panel
The backspace key is now processed in the message panel.  Echoing while
the message is typed didn't seem to be an optimal solution.  It is now
replaced by processing of each character, and if the character is not
backspace, it is displayed, while the backspace character causes the
previously displayed character to be replaced by a space and the cursor
moved back by one position.  The effect is the same as with the echo()
function, but the effect of the backspace key is immediately taken into
account when the message is populated with content.