Implement incoming call request popup
A new class CallPopup is created for displaying information about an
incoming voice call request at the end of the user that receives the
call.  The user can accept or reject the call request.  The action()
functions for accept and reject buttons are implemented, but their
output is not further processed at the moment.