Start working on handling audio
A namespace nc (for nchat) is created containing two classes:
AudioPlayer and AudioRecorder.  These will then be used to create
objects that can process audio: play packets received from the server
and record and send packets back.  The classes so far look very similar,
with the only difference being in their main functions: play() in
AudioPlayer and record() in AudioRecorder.  Perhaps a better design
solution would be to create a single class from which they will both be
derived.

The idea, for the moment, is to create either two new processes or two
new threads: one for playing and the other for recording sound.  They
would be spawned by the main client program and could receive messages
from it to enable their termination.  However, a possible better
solution is to handle audio events in the main loop of the NCWindow
run() function, which already processes all other input/output events.